Microservice helpo-ml estructurado
